Further refactoring of CVO in prep to move verify package
