+++ pkg/manifest/manifest.go
@@ -1,4 +1,4 @@
-package lib
+package manifest
 
 import (
 	"bytes"
