+++ pkg/manifest/manifest_test.go
@@ -1,4 +1,4 @@
-package lib
+package manifest
 
 import (
 	"flag"
