+++ pkg/verify/verify.go
@@ -19,8 +19,15 @@ import (
 	"sync"
 	"time"
 
+	"github.com/pkg/errors"
 	"golang.org/x/crypto/openpgp"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	coreclientsetv1 "k8s.io/client-go/kubernetes/typed/core/v1"
 	"k8s.io/klog"
+
+	"github.com/openshift/cluster-version-operator/pkg/manifest"
+	"github.com/openshift/cluster-version-operator/pkg/verify/verifyconfigmap"
 )
 
 // Interface performs verification of the provided content. The default implementation
@@ -316,6 +323,39 @@ func (v *ReleaseVerifier) Signatures() map[string][][]byte {
 		copied[k] = v
 	}
 	return copied
+}
+
+// LoadConfigMapVerifierDataFromUpdate fetches the first config map in the payload with the correct annotation.
+// It returns an error if the data is not valid, or no verifier if no config map is found. See the verify
+// package for more details on the algorithm for verification. If the annotation is set, a verifier or error
+// is always returned.
+func LoadConfigMapVerifierDataFromUpdate(manifests []manifest.Manifest, clientBuilder ClientBuilder, configMapClient coreclientsetv1.ConfigMapsGetter) (Interface, *StorePersister, error) {
+	configMapGVK := corev1.SchemeGroupVersion.WithKind("ConfigMap")
+	for _, manifest := range manifests {
+		if manifest.GVK != configMapGVK {
+			continue
+		}
+		if _, ok := manifest.Obj.GetAnnotations()[ReleaseAnnotationConfigMapVerifier]; !ok {
+			continue
+		}
+		src := fmt.Sprintf("the config map %s/%s", manifest.Obj.GetNamespace(), manifest.Obj.GetName())
+		data, _, err := unstructured.NestedStringMap(manifest.Obj.Object, "data")
+		if err != nil {
+			return nil, nil, errors.Wrapf(err, "%s is not valid: %v", src, err)
+		}
+		verifier, err := NewFromConfigMapData(src, data, clientBuilder)
+		if err != nil {
+			return nil, nil, err
+		}
+
+		// allow the verifier to consult the cluster for signature data, and also configure
+		// a process that writes signatures back to that store
+		signatureStore := verifyconfigmap.NewStore(configMapClient, nil)
+		verifier = verifier.WithStores(signatureStore)
+		persister := NewSignatureStorePersister(signatureStore, verifier)
+		return verifier, persister, nil
+	}
+	return nil, nil, nil
 }
 
 // hasVerified returns true if the digest has already been verified.
